+++ changelog.md
@@ -1,3 +1,11 @@
+## 1.1.126 *\[2021-12-??\]*
+
+#### Improvements
+
+- Add `DiscordHTTPClient.sticker_pack_get`.
+- Add `RATE_LIMIT_GROUPS.sticker_pack_get`.
+- Add `Client.sticker_pack_get`.
+
 ## 1.1.126 *\[2021-12-05\]*
 
 #### Improvements